Output json object with embedded image.
In order for the adaguc-checker WPS to be able to find the images
for each layer and to not have to clean up image files after checking
we embed the image inside json and print it to stdout so WPS can read
it and pass it on to HTML/JS frontend.

TODO: cfchecker messages are directed to /dev/null temporarily. We would
like to add those as an additional JSON item.