Catch and log exception when the ADAGUC getMap result cannot be processed.
When an incorrect image was returned by ADAGUC, the checker script failed
and did not return a JSON response. As a result, there was no report
visible in the frontend. By catching exeptions in the image handling
part, the checker script will continue and output at least the report
statements retrieved from ADAGUC.